+++ src/Qiniu/Storage/ResumeBlocker.cs
@@ -8,17 +8,17 @@ namespace Qiniu.Storage
     {
         public ManualResetEvent DoneEvent { set; get; }
         public byte[] BlockBuffer { set; get; }
-        public int BlockIndex { set; get; }
+        public long BlockIndex { set; get; }
         public string UploadToken { set; get; }
         public PutExtra PutExtra { set; get; }
         public ResumeInfo ResumeInfo { set; get; }
-        public Dictionary<int, HttpResult> BlockMakeResults;
+        public Dictionary<long, HttpResult> BlockMakeResults;
         public object ProgressLock { set; get; }
         public Dictionary<string, long> UploadedBytesDict { set; get; }
         public long FileSize { set; get; }
 
-        public ResumeBlocker(ManualResetEvent doneEvent, byte[] blockBuffer, int blockIndex, string uploadToken,
-            PutExtra putExtra, ResumeInfo resumeInfo, Dictionary<int, HttpResult> blockMakeResults,
+        public ResumeBlocker(ManualResetEvent doneEvent, byte[] blockBuffer, long blockIndex, string uploadToken,
+            PutExtra putExtra, ResumeInfo resumeInfo, Dictionary<long, HttpResult> blockMakeResults,
             object progressLock, Dictionary<string, long> uploadedBytesDict, long fileSize)
         {
             this.DoneEvent = doneEvent;
